Fix(리뷰) : 작성자 기준 리뷰 조회 유저아이디 타입 수정

+++ src/main/java/kr/ac/hongik/dsc2023/ydy/team1/core/konbini/repository/CommentRepository.java
@@ -7,6 +7,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CommentRepository extends JpaRepository<Comment,Long> {
     Page<Comment> getCommentsByPromotionInfo_Item_Id(long itemId, Pageable pageable);
-    Page<Comment> getCommentsByMember_Id(long memberID, Pageable pageable);
+    Page<Comment> getCommentsByMember_Id(int memberID, Pageable pageable);
     Page<Comment> getCommentsByPromotionInfo_Id(long promotionID, Pageable pageable);
 }
